consolidate config type
