+++ sentry_flink/sentry_flink/flink_runtime/flink_deploy.py
@@ -48,12 +48,6 @@ def main() -> int:
         type=str,
         help=("The config file"),
     )
-    parser.add_argument(
-        "--config-template",
-        type=str,
-        default="sentry_streams/deployment_config/config.json",
-        help=("The deployment configuration file template"),
-    )
 
     args = parser.parse_args()
 
@@ -86,8 +80,6 @@ def main() -> int:
             args.adapter,
             "--config",
             f"/apps/{args.config}",
-            "--config-template",
-            f"/apps/{args.config_template}",
         ]
     )
