+++ sentry_streams/sentry_streams/deployment_config/transformer_config.yaml
@@ -1,3 +1,6 @@
+# Default configuration
+# As of now, the assumption is that local adapters (like Arroyo)
+# do not require overrides of this file
 env:
   topics:
     events: events
@@ -13,14 +16,12 @@ pipeline:
           common: {
             bootstrap_servers: "localhost:9092",
           },
-          flink_overrides: {},
       },
 
         kafkasink: {
           common: {
             bootstrap_servers: "localhost:9092",
           },
-          flink_overrides: {},
         }
       }
   ]
